LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/anchor-inside-textarea.tentative-expected.txt



FAIL Scroll anchoring works on textarea assert_equals: The scroll position should be adjusted expected 310 but got 312
